Fix templates, controllers, validators

+++ system/modules/backend/templates/content/file/list.php
@@ -18,7 +18,7 @@
         </button>
         <ul class="dropdown-menu">
           <li>
-            <a data-action="delete" href="#">
+            <a data-action="delete" data-action-confirm="<?php echo $this->text('Are you sure? It cannot be undone!'); ?>" href="#">
               <?php echo $this->text('Delete from database and disk'); ?>
             </a>
           </li>
